finished major mixin refactor
