+++ pytorch_lightning/trainer/data_loading_mixin.py
@@ -1,5 +1,3 @@
-import os
-import re
 import warnings
 
 from torch.utils.data.distributed import DistributedSampler
